requirements.txt
discord.py>=2.3.0
nest-asyncio>=1.5.6
python-dotenv>=1.0.0
Flask>=2.3.2
requests>=2.31.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib